[MOD] : CardViewHolder 분리 #130
